fix(dispatch): move delegate_ownership under scripts/guardrails (#5643)
CI "No new root scripts" rejects new non-stub modules at scripts/*.py.
Colocate with primary_write_guard / worktree_containment; parents=2 for
primary-checkout root resolve.

X-Agent: grok/5643-ownership-warn
